end of lesson

+++ app/Http/Controllers/ListingController.php
@@ -36,7 +36,7 @@ class ListingController extends Controller
         if ($request->hasFile('logo')) {
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
-
+        $formFields['user_id'] = auth()->id();
     Listing::create($formFields);
 
     return redirect('/')->with('message', 'Listing created successfully!');
@@ -47,6 +47,9 @@ class ListingController extends Controller
     }
 
     public function update(Request $request, Listing $listing) {
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
         $formFields = $request->validate([
             "company" => 'required',
             "title" => 'required',
@@ -67,8 +70,15 @@ class ListingController extends Controller
     }
 
     public function destroy(Listing $listing) {
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully!');
+    }
+
+    public function manage() {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
